Traktor: File length and offsets are now 64 bit in streams.
[git-p4: depot-paths = "//depot/traktor/": change = 21523]

+++ code/Compress/Lzf/DeflateStreamLzf.cpp
@@ -30,14 +30,14 @@ public:
 		m_stream = 0;
 	}
 
-	int write(const void* block, int32_t nbytes)
+	int64_t write(const void* block, int64_t nbytes)
 	{
 		const uint8_t* top = static_cast< const uint8_t* >(block);
 		const uint8_t* ptr = static_cast< const uint8_t* >(block);
 
 		while (nbytes > 0)
 		{
-			int32_t ncopy = std::min< int32_t >(nbytes, int32_t(m_uncompressedBuffer.size() - m_uncompressedBufferCount));
+			int64_t ncopy = std::min< int64_t >(nbytes, int64_t(m_uncompressedBuffer.size() - m_uncompressedBufferCount));
 			std::memcpy(&m_uncompressedBuffer[m_uncompressedBufferCount], ptr, ncopy);
 			m_uncompressedBufferCount += ncopy;
 			ptr += ncopy;
@@ -74,7 +74,7 @@ public:
 			}
 		}
 
-		return int(ptr - top);
+		return int64_t(ptr - top);
 	}
 
 	void flush()
@@ -115,7 +115,7 @@ private:
 	Ref< IStream > m_stream;
 	AlignedVector< uint8_t > m_uncompressedBuffer;
 	AlignedVector< uint8_t > m_compressedBlock;
-	uint32_t m_uncompressedBufferCount;
+	int64_t m_uncompressedBufferCount;
 };
 
 T_IMPLEMENT_RTTI_CLASS(L"traktor.compress.DeflateStreamLzf", DeflateStreamLzf, IStream)
@@ -154,31 +154,31 @@ bool DeflateStreamLzf::canSeek() const
 	return false;
 }
 
-int DeflateStreamLzf::tell() const
+int64_t DeflateStreamLzf::tell() const
 {
 	T_FATAL_ERROR;
 	return 0;
 }
 
-int DeflateStreamLzf::available() const
+int64_t DeflateStreamLzf::available() const
 {
 	T_FATAL_ERROR;
 	return 0;
 }
 
-int DeflateStreamLzf::seek(SeekOriginType origin, int offset)
+int64_t DeflateStreamLzf::seek(SeekOriginType origin, int64_t offset)
 {
 	T_FATAL_ERROR;
 	return 0;
 }
 
-int DeflateStreamLzf::read(void* block, int nbytes)
+int64_t DeflateStreamLzf::read(void* block, int64_t nbytes)
 {
 	T_FATAL_ERROR;
 	return 0;
 }
 
-int DeflateStreamLzf::write(const void* block, int nbytes)
+int64_t DeflateStreamLzf::write(const void* block, int64_t nbytes)
 {
 	T_ASSERT (m_impl);
 	return m_impl->write(block, nbytes);
